requirements.txt
mido>=1.3.0
python-rtmidi>=1.5.0
Pillow>=10.0.0
